Separate movie and movies reducers
