+++ src/js/app.js
@@ -1,51 +1,31 @@
 
 import {BrowserRouter, Switch, Route} from 'react-router-dom';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 
-import HomePage from './components/HomePage';
-import React from 'react';
-import ErrorBoundarry from './components/ErrorBoundary';
-import FilmPage from './components/FilmPage';
 import '../styles/common.scss';
+
+import HomePage from './components/HomePage';
+import FilmPage from './components/FilmPage';
+import ErrorBoundarry from './components/ErrorBoundary';
 import reducers from './reducers/index';
 import saga from './sagas/sagas';
 
 if (process.env.NODE_ENV === 'development') {
-    console.log('Welcome to development');
+  console.log('Welcome to development');
 }
 
 if (process.env.NODE_ENV === 'production') {
-    console.log('Welcome to production');
+  console.log('Welcome to production');
 }
 
-const initialState = {
-  movies: {
-    searchBy: {
-      values: ['title', 'genres'],
-      active: 'title',
-    },
-    searchValue: '',
-    items: [],
-    sortBy: {
-      values: ['release_date', 'vote_average'],
-      active: 'release_date',
-    },
-  },
-  movie: {
-    item: null,
-    relatedMovies: {
-      criteria: 'genres',
-      items: [],
-    },
-  },
-};
+
 const sagaMiddleware = createSagaMiddleware();
 const store = createStore(
   reducers,
-  initialState,
   applyMiddleware(sagaMiddleware)
 );
 sagaMiddleware.run(saga, store.dispatch, store.getState);
@@ -62,4 +42,4 @@ ReactDOM.render((
       </BrowserRouter>
     </ErrorBoundarry>
   </Provider>
-  ), document.getElementById("container"));
+), document.getElementById("container"));
