+++ src/js/reducers/index.js
@@ -1,8 +1,9 @@
 import {combineReducers} from 'redux';
 
-import {movies, movie} from './movies';
+import movies from './movies';
+import movie from './movie';
 
 export default combineReducers({
   movies,
-  movie
+  movie,
 });
